Refactor to streamline handling of regular and zip files in FileExplorerAndEditor
